+++ src/main/java/com/jakubolejarczyk/vet_server/domain/dependent/VetDomain.java
@@ -9,7 +9,7 @@ public interface VetDomain extends BaseDomain {
     LocalDate getLicenseIssueDate();
     LocalDate getLicenseExpiryDate();
     String getSpecialization();
-    Integer getYearsOfExperience();
+    Long getYearsOfExperience();
     Long getAccountId();
     Long getOpeningHourId();
 }
